// packages/stencil-library/src/components/signin/signin-interfaces.ts
export interface SignInFormData {
  email: string
  password: string
}

export interface SignInLabels {
  title?: string
  description?: string
  emailLabel?: string
  emailPlaceholder?: string
  passwordLabel?: string
  passwordPlaceholder?: string
  showPasswordLabel?: string
  hidePasswordLabel?: string
  forgotPasswordText?: string
  signInButtonText?: string
  noAccountText?: string
  signUpButtonText?: string
}